+++ packages/tgweb/lib/tgweb/get_wrapper.mjs
@@ -1,3 +1,5 @@
+import { parseDocument } from "htmlparser2"
+
 const getWrapper = (siteData, path) => {
   const dirParts = path.split("/")
   dirParts.pop()
@@ -7,7 +9,29 @@ const getWrapper = (siteData, path) => {
     const wrapperPath = `${dir}/_wrapper.html`
 
     const wrapper = siteData.wrappers.find(wrapper => wrapper.path === wrapperPath)
-    if (wrapper) return wrapper
+
+    if (wrapper) {
+      if ("shared-wrapper" in wrapper.frontMatter.main) {
+        const sharedWrapperName = wrapper.frontMatter.main["shared-wrapper"]
+
+        const sharedWrapper =
+          siteData.sharedWrappers.find(sw =>
+            sw.path === `shared_wrappers/${sharedWrapperName}.html`
+          )
+
+        if (sharedWrapper) {
+          return sharedWrapper
+        }
+        else {
+          wrapper.dom =
+            parseDocument(`<div class='bg-error text-black m-1 py-1 px-2'>${message}</div>"`)
+
+          return wrapper
+        }
+      }
+
+      return wrapper
+    }
   }
 }